Five-section post-refactor cleanup: docs, tests, kinematic smoke
§1 Doc truth: ROADMAP TRL updates, uncomment planning Sphinx modules,
add pyro-port-remaining living backlog.

§2 Kinematic policy: render-smoke test (no committed PNG baselines);
update generator docstring and tests/README.

§3 Plan archive: move migration/gap docs to docs/plans/archive/ with
README index; add docs/plans/README.md.

§4 Test phase 2: merge RRT* into test_rrt.py, fold JAX mechanical tests
into test_mechanical*.py, add TransferFunction block test.

§5 Scratch triage: remove tests/manual/ and tests/bugs/; add
demo_native_comparison.py under examples/scripts/animation/.
