+++ docs/plans/README.md
@@ -1,0 +1,11 @@
+# Planning documents
+
+| File | Role |
+| --- | --- |
+| [pyro-port-remaining.md](pyro-port-remaining.md) | **Living backlog** — what is still missing vs pyro 2.0 |
+| [manipulator-abstraction.md](manipulator-abstraction.md) | `Manipulator` base class status + catalog rebase checklist |
+| [neural-blocks-collection.md](neural-blocks-collection.md) | Aspirational MLP / `control/neural.py` design notes |
+| [archive/](archive/) | Frozen migration and gap-checklist snapshots (historical) |
+
+Update `pyro-port-remaining.md` when landing or dropping a pyro port item.
+Archive completed tranche notes instead of growing them in place.
