+++ docs/plans/archive/README.md
@@ -1,0 +1,12 @@
+# Archived plans
+
+Frozen snapshots kept for historical context. **Do not treat as current status.**
+
+| File | Archived because |
+| --- | --- |
+| `catalog-migration-notes.md` | Pyro catalog plant migration complete |
+| `tool-migration-notes.md` | First tool tranche complete; remainder in `pyro-port-remaining.md` |
+| `pyro-port-gap-todo.md` | Master gap checklist superseded (reconciled 2026-06) |
+| `rrt-search-design.md` | RRT/RRT* implemented in `minilink/planning/search/` |
+
+Current backlog: [../pyro-port-remaining.md](../pyro-port-remaining.md).
