+++ tests/README.md
@@ -9,14 +9,17 @@ graphics frame keys, catalog equation references)—not implementation trivia or
 third-party print formatting. Prefer one parametrized or table-driven test over
 many near-duplicate files. Deep dynamics checks for a few representative plants
 live in `test_catalog_plant_contracts.py`; broad catalog smoke in
-`test_catalog_migration.py`; pixel graphics regression in
-`test_kinematic_regression.py`.
+`test_catalog_migration.py`; kinematic render smoke in
+`test_kinematic_regression.py` (manifest only — PNGs are local/gitignored).
 
 Shared fixtures: `graphics_contract_helpers.py` (draw-list resolution),
 `planning_helpers.py` (RRT holonomic obstacle scene).
 
 Benchmark **performance** helpers live under repo-root `benchmarks/`; import smoke
 only in `test_benchmark_smoke.py` (not correctness asserts).
+
+`tests/manual/` and `tests/bugs/` are removed — use `examples/scripts/` for
+smoke scripts and unittest for contracts.
 
 ## Core behavior without optional extras
